requirements.txt
requests
flask
flask_restful
gunicorn
python-dotenv
flask_cors